fix: terms
