+++ lib/marketing-slugs.ts
@@ -1,0 +1,30 @@
+/**
+ * Top-level URL slugs for pages that live in content/marketing/ (md-src/marketing/*.md).
+ * Matches marketingSource baseUrl "" in lib/source.ts.
+ */
+export const MARKETING_SLUGS = [
+  "about",
+  "careers",
+  "cn",
+  "community",
+  "cookie-policy",
+  "enterprise",
+  "find-us",
+  "imprint",
+  "jp",
+  "jp-cloud",
+  "kr",
+  "non-profit",
+  "oss-friends",
+  "press",
+  "pricing",
+  "pricing-self-host",
+  "privacy",
+  "research",
+  "startups",
+  "support",
+  "talk-to-us",
+  "terms",
+  "watch-demo",
+  "wrapped",
+] as const;
